Refactor SignedIn
